feat: Finished TFragments.03-Exercise-ClicksAndSavingState.
Each fragment's ImageView displays the next image in the list on click.
When the end of the list is reached, the image is reset to the first one
in the list. The current state of each fragment is saved so that after
rotation the current images are still being displayed.